P-51 Fix main problem
