+++ frontend/src/entities/comments/api/dto/comment.dto.ts
@@ -2,7 +2,7 @@ import { Account } from "@/entities/account/model/Account";
 
 export interface CommentDto {
     tlgId: number;
-    channelId: number;
+    peerId: number;
     postId: number;
     from: Account;
     message: string;
