+++ frontend/src/entities/comments/model/Comment.ts
@@ -3,7 +3,7 @@ import { Account } from "@/entities/account/model/Account";
 export interface Comment {
     tlgId: number;
     postId: number;
-    channelId: number;
+    peerId: number;
     from: Account;
     message: string;
     date: Date;
