+++ frontend/src/features/CommentsSerchResults/CommentsSearchResults.tsx
@@ -3,6 +3,6 @@ import { Comment } from "@/entities/comments/model/Comment";
 
 export const CommentsSearchResults = (props: { results: Comment[]; }) => {
     return (props.results.map((comment: Comment) => (
-        <CommentWidget key={comment.tlgId} {...comment} />
+        <CommentWidget key={comment.tlgId} comment={comment} showChannel showUsername />
     )));
 };
