+++ frontend/src/features/PostWidget/PostWidget.tsx
@@ -1,5 +1,5 @@
 import { Post } from "@/entities";
-import { channelApi } from "@/entities/channels";
+import { channelsApi } from "@/entities/channels";
 import { Card, CardContent, Divider, Typography } from "@mui/material";
 import { useQuery } from "@tanstack/react-query";
 import plural from 'plural-ru';
@@ -7,14 +7,14 @@ import { NavLink, useNavigate } from "react-router-dom";
 import ChevronRightIcon from '@mui/icons-material/ChevronRight';
 
 
-export const PostWidget = ({ post }: { post: Post; }) => {
+export const PostWidget = ({ post, showPostLink = true }: { post: Post; showPostLink: boolean; }) => {
 
     const navigate = useNavigate();
     const { data: channelInfo,
         isError: channelInfoIsError,
         error: channelInfoError,
         isFetched: infoIsFetched }
-        = useQuery(channelApi.channelQueries.details(post.peerId.toString()));
+        = useQuery(channelsApi.channelQueries.details(post.peerId.toString()));
 
     return (
         <Card key={post.tlgId} sx={{ width: "100%", padding: 1, boxSizing: "border-box" }}>
@@ -23,7 +23,7 @@ export const PostWidget = ({ post }: { post: Post; }) => {
                     padding: 1,
                     color: "rgb(52, 71, 103);",
                     "&:last-child": {
-                        paddingBottom: 1, // Adjust the padding bottom for the last child
+                        paddingBottom: 1,
                     },
                 }}
             >
@@ -39,30 +39,36 @@ export const PostWidget = ({ post }: { post: Post; }) => {
                 <Typography variant="body2" sx={{ marginTop: 1, lineHeight: 2 }}>
                     {post.message}
                 </Typography>
-                <Divider sx={{ marginTop: 2, width: "100%" }} />
-                <Typography variant="body1"
-                    sx={{
-                        display: "flex",
-                        justifyContent: "start",
-                        alignItems: "center",
-                        marginTop: 1
-                    }}>
-                    {
-                        (post.commentsCount ? post.commentsCount : 0) +
-                        plural((post.commentsCount ? post.commentsCount : 0),
-                            ' комментарий', ' комментария', ' комментариев')
-                    }
 
-                    <ChevronRightIcon
-                        sx={{
-                            display: "flex",
-                            justifyContent: "center",
-                            alignItems: "center",
-                            cursor: "pointer",
-                        }}
-                        onClick={() => { navigate(`/channels/${post.peerId}/posts/${post.tlgId}/comments`); }}
-                    />
-                </Typography>
+                {
+                    showPostLink &&
+                    <>
+                        <Divider sx={{ marginTop: 2, width: "100%" }} />
+                        <Typography variant="body1"
+                            sx={{
+                                display: "flex",
+                                justifyContent: "start",
+                                alignItems: "center",
+                                marginTop: 1
+                            }}>
+                            {
+                                (post.commentsCount ? post.commentsCount : 0) +
+                                plural((post.commentsCount ? post.commentsCount : 0),
+                                    ' комментарий', ' комментария', ' комментариев')
+                            }
+
+                            <ChevronRightIcon
+                                sx={{
+                                    display: "flex",
+                                    justifyContent: "center",
+                                    alignItems: "center",
+                                    cursor: "pointer",
+                                }}
+                                onClick={() => { navigate(`/channels/${post.peerId}/posts/${post.tlgId}`); }}
+                            />
+                        </Typography>
+                    </>
+                }
             </CardContent>
         </Card>);
 };
