--- frontend/src/pages/Home/Home.module.scss
@@ -1,3 +1,0 @@
-.main_container {
-    display: flex;
-}
