+++ frontend/src/pages/Home/Home.tsx
@@ -1,10 +1,7 @@
-import styles from "./Home.module.scss";
 import { Main } from "@/widgets/Main";
 
 export const Home = () => {
     return (
-        <div className={styles.main_container}>
-            <Main />
-        </div>
+        <Main />
     );
 };
