--- frontend/src/pages/PostComments/Comments.module.scss
@@ -1,24 +1,0 @@
-.loading {
-    display: grid;
-    align-items: center;
-    justify-items: center;
-    height: 100%;
-    width: 100%;
-}
-
-.post__comments {
-    display: flex;
-    flex-direction: column;
-    row-gap: 2rem;
-    align-items: left;
-    width: 100%;
-    height: 100%;
-}
-
-.comments-loading {
-    align-self: center;
-}
-
-.comments-no-data {
-    width: 50%;
-}
