--- frontend/src/pages/PostComments/index.ts
@@ -1,1 +1,0 @@
-export { PostComments } from './PostComments';
