+++ frontend/src/pages/PostPage/index.ts
@@ -1,0 +1,1 @@
+export { PostPage } from './PostPage';
