+++ frontend/src/widgets/Main/index.tsx
@@ -1,14 +1,25 @@
-import styles from "./styles.module.css";
 import { Users } from "@/features/Users";
 import Channels from "@/features/Channels";
+import { Box } from "@mui/material";
 
 export const Main = () => {
 
     return (
-        <div className={styles.info}>
+        <Box
+            sx={{
+                padding: 4,
+                display: "flex",
+                alignItems: "start",
+                gap: 2,
+                height: "100%",
+                width: "100%",
+                boxSizing: "border-box",
+            }}
+        >
             <Users />
 
             <Channels />
-        </div>
+
+        </Box>
     );
 };
