--- frontend/src/widgets/Main/styles.module.css
@@ -1,26 +1,0 @@
-.main {
-    display: flex;
-    flex: 1;
-    flex-direction: column;
-    padding-top: 1rem;
-    padding-bottom: 1rem;
-    margin-top: 1rem;
-    margin-left: 1rem;
-    margin-right: 1rem;
-    column-count: 2;
-}
-
-.info {
-    display: flex;
-    flex-wrap: wrap;
-    flex: 1 1 auto;
-    gap: 1rem;
-    justify-content: space-between;
-    column-count: 1;
-}
-
-@media screen and (min-width: 800px) {
-    .main {
-        column-count: 2;
-    }
-}
